# api/sys/v1/sysRole_router.pb.go
// Code generated by https://github.com/zhufuyi/sponge, DO NOT EDIT.

package v1

import (
	context "context"
	gin "github.com/gin-gonic/gin"
	errcode "github.com/zhufuyi/sponge/pkg/errcode"
	middleware "github.com/zhufuyi/sponge/pkg/gin/middleware"
	zap "go.uber.org/zap"
	strings "strings"
)

// import packages: strings. context. errcode. middleware. zap. gin.

type SysRoleLogicer interface {
	Create(ctx context.Context, req *CreateSysRoleRequest) (*CreateSysRoleReply, error)
	DeleteByID(ctx context.Context, req *DeleteSysRoleByIDRequest) (*DeleteSysRoleByIDReply, error)
	DeleteByIDs(ctx context.Context, req *DeleteSysRoleByIDsRequest) (*DeleteSysRoleByIDsReply, error)
	UpdateByID(ctx context.Context, req *UpdateSysRoleByIDRequest) (*UpdateSysRoleByIDReply, error)
	GetByID(ctx context.Context, req *GetSysRoleByIDRequest) (*GetSysRoleByIDReply, error)
	GetByCondition(ctx context.Context, req *GetSysRoleByConditionRequest) (*GetSysRoleByConditionReply, error)
	ListByIDs(ctx context.Context, req *ListSysRoleByIDsRequest) (*ListSysRoleByIDsReply, error)
	ListByLastID(ctx context.Context, req *ListSysRoleByLastIDRequest) (*ListSysRoleByLastIDReply, error)
	List(ctx context.Context, req *ListSysRoleRequest) (*ListSysRoleReply, error)
}

type SysRoleOption func(*sysRoleOptions)

type sysRoleOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *sysRoleOptions) apply(opts ...SysRoleOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithSysRoleHTTPResponse() SysRoleOption {
	return func(o *sysRoleOptions) {
		o.isFromRPC = false
	}
}

func WithSysRoleRPCResponse() SysRoleOption {
	return func(o *sysRoleOptions) {
		o.isFromRPC = true
	}
}

func WithSysRoleResponser(responser errcode.Responser) SysRoleOption {
	return func(o *sysRoleOptions) {
		o.responser = responser
	}
}

func WithSysRoleLogger(zapLog *zap.Logger) SysRoleOption {
	return func(o *sysRoleOptions) {
		o.zapLog = zapLog
	}
}

func WithSysRoleErrorToHTTPCode(e ...*errcode.Error) SysRoleOption {
	return func(o *sysRoleOptions) {
		o.httpErrors = e
	}
}

func WithSysRoleRPCStatusToHTTPCode(s ...*errcode.RPCStatus) SysRoleOption {
	return func(o *sysRoleOptions) {
		o.rpcStatus = s
	}
}

func WithSysRoleWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) SysRoleOption {
	return func(o *sysRoleOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterSysRoleRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic SysRoleLogicer,
	opts ...SysRoleOption) {

	o := &sysRoleOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &sysRoleRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type sysRoleRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                SysRoleLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *sysRoleRouter) register() {
	r.iRouter.Handle("POST", "/api/v1/sysRole", r.withMiddleware("POST", "/api/v1/sysRole", r.Create_0)...)
	r.iRouter.Handle("DELETE", "/api/v1/sysRole/:id", r.withMiddleware("DELETE", "/api/v1/sysRole/:id", r.DeleteByID_0)...)
	r.iRouter.Handle("POST", "/api/v1/sysRole/delete/ids", r.withMiddleware("POST", "/api/v1/sysRole/delete/ids", r.DeleteByIDs_0)...)
	r.iRouter.Handle("PUT", "/api/v1/sysRole/:id", r.withMiddleware("PUT", "/api/v1/sysRole/:id", r.UpdateByID_0)...)
	r.iRouter.Handle("GET", "/api/v1/sysRole/:id", r.withMiddleware("GET", "/api/v1/sysRole/:id", r.GetByID_0)...)
	r.iRouter.Handle("POST", "/api/v1/sysRole/condition", r.withMiddleware("POST", "/api/v1/sysRole/condition", r.GetByCondition_0)...)
	r.iRouter.Handle("POST", "/api/v1/sysRole/list/ids", r.withMiddleware("POST", "/api/v1/sysRole/list/ids", r.ListByIDs_0)...)
	r.iRouter.Handle("GET", "/api/v1/sysRole/list", r.withMiddleware("GET", "/api/v1/sysRole/list", r.ListByLastID_0)...)
	r.iRouter.Handle("POST", "/api/v1/sysRole/list", r.withMiddleware("POST", "/api/v1/sysRole/list", r.List_0)...)

}

func (r *sysRoleRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

var _ middleware.CtxKeyString

func (r *sysRoleRouter) Create_0(c *gin.Context) {
	req := &CreateSysRoleRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.Create(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysRoleRouter) DeleteByID_0(c *gin.Context) {
	req := &DeleteSysRoleByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByID(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysRoleRouter) DeleteByIDs_0(c *gin.Context) {
	req := &DeleteSysRoleByIDsRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByIDs(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysRoleRouter) UpdateByID_0(c *gin.Context) {
	req := &UpdateSysRoleByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UpdateByID(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysRoleRouter) GetByID_0(c *gin.Context) {
	req := &GetSysRoleByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByID(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysRoleRouter) GetByCondition_0(c *gin.Context) {
	req := &GetSysRoleByConditionRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByCondition(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysRoleRouter) ListByIDs_0(c *gin.Context) {
	req := &ListSysRoleByIDsRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListByIDs(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysRoleRouter) ListByLastID_0(c *gin.Context) {
	req := &ListSysRoleByLastIDRequest{}
	var err error

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListByLastID(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysRoleRouter) List_0(c *gin.Context) {
	req := &ListSysRoleRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.List(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}